// src/interfaces/report.ts
export interface Report {
  id: string;
  reportNumber: string;
  date: string;
  teamNumber: string;
  resultVial1: string;
  resultVial2: string;
  resultVial3: string;
  comments: string;

  sterilizerBrand: string;
  sterilizerModel: string;
  sterilizerSerialNumber: string;
  sterilizerSterilizerType: string;
}